Put field converter to separate file
